aws_devops/pyproject.toml
[tool.poetry]
name = "label-studio-demo"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"

# AWS CLI and dependencies
awscli = "1.37.11"
botocore = "1.36.11"
docutils = "<0.17,>=0.10"
s3transfer = "0.11.2"
PyYAML = "6.0.2"
colorama = "0.4.6"
rsa = "<4.8,>=3.1.2"
jmespath = "1.0.1"
python-dateutil = "<3.0.0,>=2.1"
urllib3 = "!=2.2.0,<3,>=1.25.4"
pyasn1 = ">=0.1.3"
six = "1.17.0"

# AWS CDK dependencies
aws-cdk-lib = "2.173.4"
constructs = "10.4.2"
typeguard = "2.13.3"
jsii = "1.106.0"
publication = "0.0.3"
attrs = "<25.0,>=21.2"
cattrs = "<24.2,>=1.8"
importlib-resources = ">=5.2.0"
typing-extensions = "<5.0,>=3.8"
aws-cdk-aws-glue-alpha = "2.159.1a0"
boto3 = "1.36.11"
pyspark = "3.3.0"
awsglue3-local = "1.0.0"
label-studio = "1.15.0"
label-studio-sdk = "1.0.8"
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.7"
pre-commit = "^3.8.0"
pytest = "^7.4.4"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
moto = "^5.0.14"

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true

[tool.ruff.lint]
select = [
    "F", "E", "W", "I", "N", "UP", "YTT", "S102", "B", "C4", "T10", "FA",
    "ISC", "ICN", "LOG", "G", "PIE", "PYI", "PT", "Q", "RSE", "SLOT",
    "TID", "TCH", "INT", "ARG", "PGH", "PL", "FLY", "NPY", "PERF", "RUF"
]
ignore = [
    "G004", "PLR0913", "E501"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
